+++ doc/changes.rst
@@ -6,8 +6,10 @@ desispec Change Log
 -------------------
 
 * Pipeline fix to allow redrock to use a full node per healpix (PR `#585`_).
+* Update pipeline maxtime/maxnodes job calculation (PR `#588`_).
 
 .. _`#585`: https://github.com/desihub/desispec/pull/585
+.. _`#588`: https://github.com/desihub/desispec/pull/588
 
 0.20.0 (2018-03-29)
 -------------------